Переносит события, не связанные с масштабом, в точку входа
